+++ library/core/designsystem/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/core/designsystem/ui/icons/ToolkitIcon.kt
@@ -97,6 +97,6 @@ sealed interface ToolkitIcon {
             @DrawableRes resId: Int,
             atEnd: Boolean = false,
             replayMode: ToolkitIconReplayMode = ToolkitIconReplayMode.Restart,
-        ): ToolkitIcon = AnimatedVector(resId, atEnd, replayMode)
+        ): AnimatedVector = AnimatedVector(resId, atEnd, replayMode)
     }
 }